Домашнее задание к уроку 8 PQL, Spring ORM, DAO на основе Spring ORM + JPA Исправление замечаний в PR

+++ src/main/java/ru/lilitweb/books/repostory/BookRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface BookRepository {
     void insert(Book book);
     void update(Book book);
-    Book getById(int id);
+    Book getById(long id);
 
     List<Book> getAll();
