+++ src/main/java/ru/lilitweb/books/repostory/BookRepositoryJpa.java
@@ -31,7 +31,7 @@ public class BookRepositoryJpa implements BookRepository {
 
 
     @Override
-    public Book getById(int id) {
+    public Book getById(long id) {
         return em.find(Book.class, id);
     }
