+++ src/main/java/ru/lilitweb/books/repostory/CommentRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface CommentRepository {
     void insert(Comment comment);
     void update(Comment comment);
-    Comment getById(int id);
+    Comment getById(long id);
 
     List<Comment> getAllByBook(Book book);
